Refactor and add SecureField
